fwk/listener/requirements.txt
aiohttp
aiodns
ccxt==4.4.90
jinja2
msgpack
numpy
pyarrow
pycares
pydantic
python-dotenv
pyyaml
pandas
zstandard
redis
uvloop
uvicorn
pytest
pytest-asyncio
fakeredis
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
PyJWT==2.8.0
python-multipart==0.0.6
PyYAML==6.0.1
APScheduler==3.10.4
httpx==0.25.2
requests
yfinance
